+++ src/components/menu-category/container/form/RemoveCategory.tsx
@@ -1,7 +1,7 @@
 import styles from '@/components/menu-category/styles/category.module.css';
+import { MENU_TOAST } from '@/data/menu-item';
 import useSetCategories from '@/hooks/query/menu/menu-category/useSetCategories';
 import useMenuToast from '@/hooks/service/menu/useMenuToast';
-import { MENU_TOAST } from '@/hooks/service/menu/useText';
 import { useModal } from '@/hooks/service/ui/useModal';
 import useCategoriesStore, { setCategory } from '@/shared/store/menu/menu-category';
 import { Tables } from '@/types/supabase';
