+++ src/components/menu-item/options/MenuOptionModalInput.tsx
@@ -1,5 +1,5 @@
+import { MENU_OPTION, MENU_TOAST } from '@/data/menu-item';
 import useMenuToast from '@/hooks/service/menu/useMenuToast';
-import { MENU_OPTION, MENU_TOAST } from '@/hooks/service/menu/useText';
 import useMenuOptionStore, {
   NewOptionDetailType,
   setMenuOption,
